# api.yaml
openapi: 3.0.3
info:
  version: 0.0.3
  title: OgiriGenerator API
paths:
  /users:
    get:
      operationId: GetUser
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        default:
          description: デフォルトのエラーレスポンス
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultErrorResponse"
    post:
      operationId: PostUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: OK
        default:
          description: デフォルトのエラーレスポンス
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultErrorResponse"
  /answers:
    post:
      operationId: PostAnswer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Answer"
      responses:
        "200":
          description: OK
        default:
          description: デフォルトのエラーレスポンス
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultErrorResponse"
  /answers/byUserId:
    get:
      operationId: GetAnswersByUserId
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Answer"
        default:
          description: デフォルトのエラーレスポンス
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultErrorResponse"
  /answers/byAnswerId:
    get:
      operationId: GetAnswerByAnswerId
      parameters:
        - name: answerId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Answer"
        default:
          description: デフォルトのエラーレスポンス
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultErrorResponse"
  /questions:
    get:
      operationId: GetQuestion
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        default:
          description: デフォルトのエラーレスポンス
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultErrorResponse"
  /questions/generate:
    post:
      operationId: GenerateQuestion
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        default:
          description: デフォルトのエラーレスポンス
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultErrorResponse"
components:
  schemas:
    User:
      type: object
      properties:
        userId:
          type: string
      required:
        - userId
    Answer:
      type: object
      properties:
        answerId:
          type: integer
        userId:
          type: string
        questionId:
          type: integer
        question:
          type: string
        answer:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - answerId
        - userId
        - questionId
        - question
        - answer
        - timestamp
    Question:
      type: object
      properties:
        questionId:
          type: integer
        question:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - questionId
        - question
        - timestamp
    DefaultErrorResponse:
      type: object
      properties:
        message:
          type: string